requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cfgv==3.4.0
chardet==5.2.0
cssselect2==0.7.0
distlib==0.3.8
Django==4.2.13
django-appconf==1.0.6
django-filer==3.1.3
django-imagekit==5.0.0
django-js-asset==2.2.0
django-polymorphic==3.1.0
django-tinymce==4.1.0
easy-thumbnails==2.8.5
filelock==3.15.4
flake8==7.1.0
identify==2.5.36
lxml==5.2.2
mccabe==0.7.0
nodeenv==1.9.1
pilkit==3.0
pillow==10.4.0
pillow_heif==0.17.0
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
reportlab==4.2.2
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
virtualenv==20.26.3
webencodings==0.5.1
